requirements.txt
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
flasgger==0.9.7.1
Flask==3.1.0
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
mistune==3.1.3
numpy==2.2.5
outcome==1.3.0.post0
packaging==25.0
pandas==2.2.3
PyJWT==2.10.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
selenium==4.32.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
